Modified script to ad link for web and check if bootstrapped

+++ boot/bootstrap.sh
@@ -2,6 +2,11 @@
 
 # required to install CHEF !!
 #
+if [ -e bootstrapped ] 
+then 
+  echo 'Flagged as already boot strapped. Exiting'
+  exit 0
+fi
  
 yum install -y gcc automake autoconf libtool make
 yum install -y ruby
@@ -13,4 +18,13 @@ gem install chef ruby-shadow --no-ri --no-rdoc
  
 #init git
 yum install -y git-core
- 
+
+#Create link to web directory
+if [ ! -h /var/www/html/www ] 
+then 
+  echo 'Creating Link to /vagrant/www'
+  ln -s /vagrant/www /var/www/html/
+fi
+
+# Touch flag 
+touch bootstrapped
